requirements.txt
Django
gunicorn
dj-database-url
django-crispy-forms
crispy-bootstrap5
whitenoise
pytz
Pillow
django-fernet-fields-v2
